[Refactor][Ops] Select a kernel and ask it how it is built (#2146)
## Problems

- Two verbs reached the kernel cache: `kernel_for` asked the selected
class how it is built, `get_or_build_kernel(key=, build=)` took a
hand-written recipe. 42 op files used the second.
- Roles were not roles. `_LogicalReduceOpBase`, `AvgPool*FwdOp` and
`MaxPool*FwdOp` keyed memoization on the selected kernel's name;
`GroupedQueryAttentionBwdOp` took two roles for one forward;
`MoePermuteAlignFwdOp` built its kernel in `__init__`, reaching no cache
at all.
- Every compile boundary was hand-written — 99 `custom_op`
registrations, 95 `register_fake` bodies, four per-family factories —
each restating its manifest entry, so a schema could fall behind the
entry with nothing to catch it.
- 41 implemented ops registered their operator in a kernel file and
declared no `torch_compile_fullgraph`, so the node a compiled graph
carried was one kernel configuration.
- An output dtype had no single origin: the manifest declared a union
where the caller decides, `_fake_output_dtype` let an op answer for its
fake, and three ops restated in code which input the fallback follows. A
fake could report one dtype while its op returned another.
- `GemmFp8FwdOp` declared `bias` as `bfloat16` while requiring it to
match `out_dtype`, which the same entry lets a caller set to `float16`.
- `TopkSelectorKernel` never wrote the slots a window shorter than
`topk` leaves unselected, returning whatever the buffer held; two
identical calls disagreed on 6 of 64.
- Four ops took a manifest `param` as a call argument.

## Changes

- `Op.kernel_for(role, inputs, call)` is the only way an op reaches a
kernel; `Op.entry_for(role, call)` answers with the identity and the
builder. The default selects among the op's candidates and asks the
chosen class; an op with one implementation overrides it. All 42 op
files migrated, roles are one per op, and `GroupedQueryAttentionBwdOp`
holds both kernels in one entry.
- New `_compile_boundary_codegen` generates each operator and fake from
the manifest entry, run from `Op.__init_subclass__`. An op declares one
`OperatorSpec` per operator, stating which of three kinds it is —
returning the declared outputs, writing into an argument, writing a
caller-supplied buffer — and which argument the writing kinds write.
**40 registrations under `src/tileops/ops/` and 59 under
`src/tileops/kernels/` become 0 and 0; 95 `register_fake` bodies become
the one the codegen calls.** `reduction/_boundary.py` is deleted
outright.
- 86 ops declare the boundary and split `forward` into one operator call
plus `_eager_forward`. `Mamba2FwdOp` stays off it as a composite whose
graph holds its leaves' operators, `DeltaNetAutogradOp` because it needs
`register_autograd`.
- Output dtype has one origin per output: an output the caller may
restate declares `caller_stated: true` and reads a param named
`out_dtype`, every other keeps what the entry declares, and a
declaration resolves to one dtype rather than a set.
`_fake_output_dtype` and `RoutingEpilogueSpec.output_dtype` are deleted;
`_output_dtype.output_dtype` is the one implementation the fake, the
three eager paths and the roofline share; `manifest.md` R23 states the
rule and `check_c9_output_dtype_convention` blocks on it in both
directions under `--strict`.
- `GemmFp8FwdOp`'s `bias` is `same_as(d)`, which is what the op
enforces. The declaration could not be written while the output was
`bfloat16 | float16`.
- `TopkSelectorKernel` writes the unselected slots with `seq_len_kv`,
one past the last key, which every consumer of an index list treats as
selecting nothing. That the slot order is unspecified, unlike the sorted
indices of the `ref_api` it names, is now on the op and in the entry.
- The operator name is `tileops::<family>_<snake(class)>`, family named
once: `norm_*` becomes `normalization_*`, `conv_*` becomes
`convolution_*`. No op chooses its own name.
- `IndexedExpertMLPFwdOp` (merged in #2141) called the deleted
`get_or_build_kernel`, so it could not have run as merged. It reaches
its kernels and its boundary like every other op now, and gets the
manifest entry it had none of.
- One compile-contract module, `tests/ops/test_compile_boundary.py`, one
`_<family>_cases()` per family. `assert_fake_matches_eager` holds each
fake to what its op returns; `TestCompileContractRegistry` holds the
manifest's `torch_compile_fullgraph` set equal to the registered
evidence. An op-only change now triggers the manifest job, because the
parity gates read the op classes.
- `manifest.md` gains R23, drops the generated `dtype` ctor block that
no longer exists, stops permitting a union on an output (R5), and names
`dtype` the reserved workload key a `roofline.func` formula reads (R21).
`.claude/domain-rules/manifest-spec.md` drops the rule that a new entry
must land `spec-only`.

## Public API changes

Four ops took a declared `param` per call; all four are construction
arguments now. `MHCPreFwdOp` would have raised `AttributeError` from
`_manifest_params()` the first time an external target served it; the
paged prefill op held a `self.max_seqlen_q` its constructor never set.

| op | moved to `__init__` |
| --- | --- |
| `MHCPreFwdOp` | `alpha_pre`, `alpha_post`, `alpha_res`,
`sinkhorn_repeat`, `sinkhorn_eps` |
| `GLABwdOp` | `has_initial_state` |
| `GroupedQueryAttentionSlidingWindowVarlenFwdOp` | `max_seqlen_q` |
| `GroupedQueryAttentionPrefillPagedWithKVCacheFwdOp` | `max_seqlen_q` |

Six ops rename a construction parameter to `out_dtype`, the one name R23
gives a caller-stated output dtype: `AlibiFwdOp` and `SinusoidalFwdOp`
(which also gain a `float32` default), `DaCumsumFwdOp`, and the three
GQA ops that had it as `dtype`. `GemmFp8FwdOp` and `BmmFp8FwdOp` keep
the name and take a `torch.dtype` rather than a dtype spelled as a
string.

## Found, not fixed

Both stay in `_NONDETERMINISTIC`, so a compiled call is held to the
traced graph plus the declared shapes and dtypes rather than to
equality.

- `SparseMlaKernel` (`DeepSeekSparseAttentionDecodeWithKVCacheFwdOp`)
reads storage that is not one of its inputs. Two identical eager calls
agree exactly, but a NaN-filled caching allocator between them turns the
whole second output to NaN — 2097152 of 2097152 — with every floating
input still clean. Deterministic and poison-sensitive together means the
NaN arrives through a read, not an unwritten slot. Not caused by this
change.
- `TopkSelectorFwdOp` returns a set rather than a sequence: it selects
the same positions every call and lets the atomic increments that claim
the slots decide which index lands where.

## Regression

One kernel body changed, `TopkSelectorKernel`, whose own test could not
see the unwritten slots because its workloads select over the full key
range. `check_c8_mutated_inputs_parity` holds the union of an op's
operators' write arguments equal to the manifest's `mutated: true`
inputs. Each written argument now carries its own alias symbol; sharing
one said two arguments may be the same tensor.

+++ benchmarks/ops/bench_independent_elementwise.py
@@ -151,7 +151,7 @@ def _sinusoidal_reference(seq_len: int, d_model: int, dtype: torch.dtype) -> tor
 
 @AlibiBenchFixture
 def test_alibi_bench(seq_len: int, num_heads: int, dtype: torch.dtype) -> None:
-    op = AlibiFwdOp(seq_len=seq_len, num_heads=num_heads, dtype=dtype)
+    op = AlibiFwdOp(seq_len=seq_len, num_heads=num_heads, out_dtype=dtype)
     workload = _GenerativeWorkload((num_heads, seq_len, seq_len), dtype)
     bm = ManifestBenchmark(op, workload)
 
@@ -169,7 +169,7 @@ def test_alibi_bench(seq_len: int, num_heads: int, dtype: torch.dtype) -> None:
 
 @SinusoidalBenchFixture
 def test_sinusoidal_bench(seq_len: int, d_model: int, dtype: torch.dtype) -> None:
-    op = SinusoidalFwdOp(seq_len=seq_len, d_model=d_model, dtype=dtype)
+    op = SinusoidalFwdOp(seq_len=seq_len, d_model=d_model, out_dtype=dtype)
     workload = _GenerativeWorkload((seq_len, d_model), dtype)
     bm = ManifestBenchmark(op, workload)
